+++ linux/flutter/generated_plugin_registrant.cc
@@ -8,6 +8,7 @@
 
 #include <bluetooth_low_energy_linux/bluetooth_low_energy_linux_plugin.h>
 #include <file_selector_linux/file_selector_plugin.h>
+#include <flutter_screen_capture/flutter_screen_capture_plugin.h>
 #include <flutter_secure_storage_linux/flutter_secure_storage_linux_plugin.h>
 #include <flutter_volume_controller/flutter_volume_controller_plugin.h>
 #include <gamepads_linux/gamepads_linux_plugin.h>
@@ -25,6 +26,9 @@ void fl_register_plugins(FlPluginRegistry* registry) {
   g_autoptr(FlPluginRegistrar) file_selector_linux_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "FileSelectorPlugin");
   file_selector_plugin_register_with_registrar(file_selector_linux_registrar);
+  g_autoptr(FlPluginRegistrar) flutter_screen_capture_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "FlutterScreenCapturePlugin");
+  flutter_screen_capture_plugin_register_with_registrar(flutter_screen_capture_registrar);
   g_autoptr(FlPluginRegistrar) flutter_secure_storage_linux_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "FlutterSecureStorageLinuxPlugin");
   flutter_secure_storage_linux_plugin_register_with_registrar(flutter_secure_storage_linux_registrar);